Made the undo stack larger (320 moves). Made the cursor exit notes mode when undoing.

+++ game/game_state.h
@@ -28,9 +28,8 @@ struct GameState_t
 	
 	Board_t board;
 	Cursor_t cursor;
-	u8 undoIndex;
+	u16 undoIndex;
 	MoveList_t history;
-	
 	
 	bool screenIsDirty;
 	bool nextUpdateIsDirty;
